getSpecWrapper util for locating DOM components, per EventBright best practices doc
